Crate turbopack_static

Static asset support for turbopack.

Static assets are copied directly to the output folder.

When imported from ES modules, they produce a thin module that simply exports the asset’s path.

When referred to from CSS assets, the reference is replaced with the asset’s path.

Structs

StaticModuleAsset

StaticModuleAssetVc

Vc for StaticModuleAsset

Functions

register

Type Definitions

StaticModuleAssetReadRef

see [turbo_tasks::ReadRef]